requirements.txt
streamlit==1.51.0
fastapi==0.118.3
uvicorn==0.27.1
requests==2.31.0
adk==0.0.5
google-ai-generativelanguage==0.4.0
python-dotenv==1.0.1
python-multipart==0.0.9
typing-extensions==4.9.0